mcp_server/pyproject.toml
[project]
name = "mcp_server"
version = "0.1.0"
description = "MCP Server providing search_web via Streamable HTTP"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mcp>=1.10.0",
    "tavily-python>=0.7.11",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
mcp-server = "mcp_server.server:main"
devmate-mcp-server = "mcp_server.server:main"

[tool.hatch.build.targets.wheel]
packages = ["src/mcp_server"]
